# swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "This is a key value store server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
      "version": "1.0.0",
      "title": "Swagger Petstore",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
         "email": "apiteam@swagger.io"
      },
      "license": {
         "name": "Apache 2.0",
         "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
   "host": "mydoc-challenge.herokuapp.com",
   "basePath": "",
   "tags": [
      {
         "name": "Key Value Store",
         "description": "Everything about key value store server"
      }
   ],
   "schemes": [
      "https",
      "http"
   ],
   "paths": {
      "/object/{key}": {
         "get": {
            "tags": [
               "Key Value Store"
            ],
            "summary": "get value of key",
            "description": "",
            "operationId": "getKeyValue",
            "consumes": [
               "application/json"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "path",
                  "name": "key",
                  "description": "key data was stored",
                  "required":"true",
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "in": "query",
                  "name": "timestamp",
                  "description": "timestamp at which value was stored",
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "get latest value of key",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "value": {
                           "type": "string",
                           "example": "value1"
                        }
                     }
                  }
               }
            },
            "security": [
               {
                  "driver_auth": [
                     "write:vehicle",
                     "read:vehicle"
                  ]
               }
            ]
         }
      },
      "/object": {
         "post": {
            "tags": [
               "Key Value Store"
            ],
            "summary": "create key value store",
            "description": "",
            "operationId": "postKeyValue",
            "consumes": [
               "application/json"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "body",
                  "name": "body",
                  "description": "key value data",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "key": {
                           "type": "string",
                           "example":"value1"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "list vehicle types",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "key": {
                           "type": "string",
                           "example": "mykey"
                        },
                        "value": {
                           "type": "string",
                           "example": "value1"
                        },
                        "timestamp": {
                           "type": "integer",
                           "example": 12345678
                        }
                     }
                  }
               }
            },
            "security": [
               {
                  "driver_auth": [
                     "write:vehicle",
                     "read:vehicle"
                  ]
               }
            ]
         }
      }
   },
   "securityDefinitions": {
      "petstore_auth": {
         "type": "oauth2",
         "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
         "flow": "implicit",
         "scopes": {
            "write:pets": "modify pets in your account",
            "read:pets": "read your pets"
         }
      },
      "api_key": {
         "type": "apiKey",
         "name": "api_key",
         "in": "header"
      }
   },
   "definitions": {
   },
   "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
   }
}